+++ libs/wire-api-federation/src/Wire/API/Federation/Client.hs
@@ -36,6 +36,9 @@ import Util.Options (Endpoint (..))
 import Wire.API.Federation.GRPC.Client (createGrpcClient, reason)
 import qualified Wire.API.Federation.GRPC.Types as Proto
 
+-- FUTUREWORK: Remove originDomain from here and make it part of all the API
+-- calls or figure out some smarter way of making API calls so it doesn't have
+-- to be specified everywhere.
 data FederatorClientEnv = FederatorClientEnv
   { grpcClient :: GrpcClient,
     targetDomain :: Domain,