+++ services/galley/test/integration/API/Federation.hs
@@ -22,12 +22,19 @@ import Bilge
 import Bilge.Assert
 import qualified Cassandra as Cql
 import Control.Lens hiding ((#))
+import Data.Aeson (ToJSON (..))
+import qualified Data.Aeson as A
+import Data.ByteString.Conversion (toByteString')
+import qualified Data.ByteString.Lazy as LBS
 import Data.Domain
-import Data.Id (Id (..), newClientId, randomId)
+import Data.Id (ConvId, Id (..), newClientId, randomId)
+import Data.Json.Util (Base64ByteString (..), toBase64Text)
 import Data.List1
 import qualified Data.List1 as List1
 import qualified Data.Map as Map
+import qualified Data.ProtoLens as Protolens
 import Data.Qualified (Qualified (..))
+import qualified Data.Set as Set
 import Data.Time.Clock
 import Data.Timeout (TimeoutUnit (..), (#))
 import Data.UUID.V4 (nextRandom)
@@ -35,6 +42,7 @@ import qualified Galley.Data.Queries as Cql
 import Galley.Types
 import Gundeck.Types.Notification
 import Imports
+import Test.QuickCheck (arbitrary, generate)
 import Test.Tasty
 import qualified Test.Tasty.Cannon as WS
 import Test.Tasty.HUnit
@@ -43,6 +51,10 @@ import TestSetup
 import Wire.API.Conversation.Role
 import Wire.API.Federation.API.Galley (GetConversationsRequest (..), GetConversationsResponse (..))
 import qualified Wire.API.Federation.API.Galley as FedGalley
+import qualified Wire.API.Federation.GRPC.Types as F
+import Wire.API.Message (ClientMismatchStrategy (..), MessageSendingStatus (mssDeletedClients, mssFailedToSend, mssRedundantClients), mkQualifiedOtrPayload, mssMissingClients)
+import Wire.API.User.Client (PubClient (..))
+import Wire.API.User.Profile
 
 tests :: IO TestSetup -> TestTree
 tests s =
@@ -58,7 +70,8 @@ tests s =
         s
         "POST /federation/update-conversation-memberships : Notify local user about other members joining"
         notifyLocalUser,
-      test s "POST /federation/receive-message : Receive a message from another backend" receiveMessage
+      test s "POST /federation/receive-message : Receive a message from another backend" receiveMessage,
+      test s "POST /federation/send-message : Post a message sent from another backend" sendMessage
     ]
 
 getConversationsAllFound :: TestM ()
@@ -250,3 +263,111 @@ receiveMessage = do
           evtData e @?= EdOtrMessage (OtrMessage fromc alicec txt Nothing)
       -- eve should not receive the message
       WS.assertNoEvent (1 # Second) [wsE]
+
+-- alice local, bob and chad remote in a local conversation
+-- bob sends a message (using the RPC), we test that alice receives it and that
+-- a call is made to the receiveMessage RPC to inform chad
+sendMessage :: HasCallStack => TestM ()
+sendMessage = do
+  cannon <- view tsCannon
+  let remoteDomain = Domain "far-away.example.com"
+  localDomain <- viewFederationDomain
+
+  -- users and clients
+  (alice, aliceClient) <- randomUserWithClientQualified (someLastPrekeys !! 0)
+  let aliceId = qUnqualified alice
+  bobId <- randomId
+  bobClient <- liftIO $ generate arbitrary
+  let bob = Qualified bobId remoteDomain
+      bobProfile = mkProfile bob (Name "Bob")
+  chadId <- randomId
+  chadClient <- liftIO $ generate arbitrary
+  let chad = Qualified chadId remoteDomain
+      chadProfile = mkProfile chad (Name "Chad")
+
+  -- conversation
+  opts <- view tsGConf
+  let responses1 req
+        | fmap F.component (F.request req) == Just F.Brig =
+          toJSON [bobProfile, chadProfile]
+        | otherwise = toJSON ()
+  (convId, requests1) <-
+    withTempMockFederator opts remoteDomain responses1 $
+      fmap decodeConvId $
+        postConvQualified aliceId [bob, chad] Nothing [] Nothing Nothing
+          <!! const 201 === statusCode
+
+  liftIO $ do
+    [brigReq, galleyReq] <- case requests1 of
+      xs@[_, _] -> pure xs
+      _ -> assertFailure "unexpected number of requests"
+    fmap F.component (F.request brigReq) @?= Just F.Brig
+    fmap F.path (F.request brigReq) @?= Just "/federation/get-users-by-ids"
+    fmap F.component (F.request galleyReq) @?= Just F.Galley
+    fmap F.path (F.request galleyReq) @?= Just "/federation/register-conversation"
+  let conv = Qualified convId localDomain
+
+  -- we use bilge instead of the federation client to make a federated request
+  -- here, because we need to make use of the mock federator, which at the moment
+  -- supports only bilge requests
+  let rcpts =
+        [ (alice, aliceClient, "hi alice"),
+          (chad, chadClient, "hi chad")
+        ]
+      msg = mkQualifiedOtrPayload bobClient rcpts "" MismatchReportAll
+      msr =
+        FedGalley.MessageSendRequest
+          { FedGalley.msrConvId = convId,
+            FedGalley.msrSender = bobId,
+            FedGalley.msrRawMessage =
+              Base64ByteString
+                (LBS.fromStrict (Protolens.encodeMessage msg))
+          }
+  let responses2 req
+        | fmap F.component (F.request req) == Just F.Brig =
+          toJSON
+            ( Map.fromList
+                [ (chadId, Set.singleton (PubClient chadClient Nothing)),
+                  (bobId, Set.singleton (PubClient bobClient Nothing))
+                ]
+            )
+        | otherwise = toJSON ()
+  (_, requests2) <- withTempMockFederator opts remoteDomain responses2 $ do
+    WS.bracketR cannon aliceId $ \ws -> do
+      g <- viewGalley
+      msresp <-
+        post
+          ( g
+              . paths ["federation", "send-message"]
+              . content "application/json"
+              . header "Wire-Origin-Domain" (toByteString' remoteDomain)
+              . json msr
+          )
+          <!! do
+            const 200 === statusCode
+      (FedGalley.MessageSendResponse eithStatus) <- responseJsonError msresp
+      liftIO $ case eithStatus of
+        Left err -> assertFailure $ "Expected Right, got Left: " <> show err
+        Right mss -> do
+          assertEqual "missing clients should be empty" mempty (mssMissingClients mss)
+          assertEqual "redundant clients should be empty" mempty (mssRedundantClients mss)
+          assertEqual "deleted clients should be empty" mempty (mssDeletedClients mss)
+          assertEqual "failed to send should be empty" mempty (mssFailedToSend mss)
+
+      -- check that alice received the message
+      WS.assertMatch_ (5 # Second) ws $
+        wsAssertOtr' "" conv bob bobClient aliceClient (toBase64Text "hi alice")
+
+  -- check that a request to propagate message to chad has been made
+  liftIO $ do
+    [_clientReq, receiveReq] <- case requests2 of
+      xs@[_, _] -> pure xs
+      _ -> assertFailure "unexpected number of requests"
+    fmap F.component (F.request receiveReq) @?= Just F.Galley
+    fmap F.path (F.request receiveReq) @?= Just "/federation/receive-message"
+    rm <- case A.decode . LBS.fromStrict . F.body =<< F.request receiveReq of
+      Nothing -> assertFailure "invalid federated request body"
+      Just x -> pure (x :: FedGalley.RemoteMessage ConvId)
+    FedGalley.rmSender rm @?= bob
+    Map.keysSet (userClientMap (FedGalley.rmRecipients rm))
+      @?= Set.singleton chadId