+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * [Federation] RPC to propagate messages to other backends (#1596).
 * [Federation] Fetch remote user's clients when sending messages (#1635).
 * [Federation] Actually propagate messages to other backends (#1638).
+* [Federation] Support sending messages to remote conversations (#1609).
 
 ## Internal changes